Adding as many icons as there's different build causes.
The previous code didn't handle cases where there would be
many causes for the build.
Sure, most of the time there's only one cause. But having multiple
one happens. For example: if an upstream project was just built and
at that moment Jenkins triggers a build because of timer trigger.