Build with JDK11 only on recent enough Jenkins
Note:
```
+  [ platform: "linux", jdk: "8", jenkins: coreJdk11Version, javaLevel: "8" ],
+  [ platform: "windows", jdk: "8", jenkins: coreJdk11Version, javaLevel: "8" ],
```

is not strictly necessary but is nice to have to check the plugin builds
correctly with JDK8 on the given recent core.
This way, if this fails on JDK11, we know it's only because of the JDK
bump, and not because of a Jenkins core change.

Similar to https://github.com/jenkinsci/label-verifier-plugin/pull/4